rfctr(part): extract partition.common submodules
In preparation for consolidating post-partitioning metadata decorators,
extract `partition.common` module into a sub-package (directory) and
extract `partition.common.metadata` module to house metadata-specific
object shared by partitioners.

This new module will be the home of the new consolidated metadata
decorator.
