+++ unstructured/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "0.15.13"  # pragma: no cover
+__version__ = "0.15.14-dev0"  # pragma: no cover
