+++ unstructured/partition/api.py
@@ -9,7 +9,7 @@ from unstructured_client.models import shared
 
 from unstructured.documents.elements import Element
 from unstructured.logger import logger
-from unstructured.partition.common import exactly_one
+from unstructured.partition.common.common import exactly_one
 from unstructured.staging.base import elements_from_dicts, elements_from_json
