+++ unstructured/partition/common/common.py
@@ -1,9 +1,7 @@
 from __future__ import annotations
 
 import numbers
-import os
 import subprocess
-from datetime import datetime
 from io import BufferedReader, BytesIO, TextIOWrapper
 from tempfile import SpooledTemporaryFile
 from time import sleep
@@ -40,76 +38,6 @@ if dependency_exists("numpy") and dependency_exists("cv2"):
 if TYPE_CHECKING:
     from unstructured_inference.inference.layout import DocumentLayout, PageLayout
     from unstructured_inference.inference.layoutelement import LayoutElement
-
-HIERARCHY_RULE_SET = {
-    "Title": [
-        "Text",
-        "UncategorizedText",
-        "NarrativeText",
-        "ListItem",
-        "BulletedText",
-        "Table",
-        "FigureCaption",
-        "CheckBox",
-        "Table",
-    ],
-    "Header": [
-        "Title",
-        "Text",
-        "UncategorizedText",
-        "NarrativeText",
-        "ListItem",
-        "BulletedText",
-        "Table",
-        "FigureCaption",
-        "CheckBox",
-        "Table",
-    ],
-}
-
-
-def get_last_modified(
-    filename: str | None, file: IO[bytes] | None, date_from_file_object: bool
-) -> str | None:
-    """Determine best available last-modified date from file or filename."""
-    if filename is not None:
-        return get_last_modified_date(filename)
-
-    if file is not None:
-        return get_last_modified_date_from_file(file) if date_from_file_object else None
-
-    return None
-
-
-def get_last_modified_date(filename: str) -> Optional[str]:
-    """Modification time of file at path `filename`, if it exists.
-
-    Returns `None` when `filename` is not a path to a file on the local filesystem.
-
-    Otherwise returns date and time in ISO 8601 string format (YYYY-MM-DDTHH:MM:SS) like
-    "2024-03-05T17:02:53".
-    """
-    if not os.path.isfile(filename):
-        return None
-
-    modify_date = datetime.fromtimestamp(os.path.getmtime(filename))
-    return modify_date.strftime("%Y-%m-%dT%H:%M:%S%z")
-
-
-def get_last_modified_date_from_file(file: IO[bytes] | bytes) -> Optional[str]:
-    """Modified timestamp of `file` if it corresponds to a file on the local filesystem."""
-    # -- a file-like object will have a name attribute if created by `open()` or if a name is
-    # -- assigned to it for metadata purposes. Use "" as default because the empty string is never
-    # -- a path to an actual file.
-    filename = str(getattr(file, "name", ""))
-
-    # -- there's no guarantee the path corresponds to an actual file on the filesystem. In
-    # -- particular, a user can set the `.name` attribute of an e.g. `io.BytesIO` object to
-    # -- populate the `.metadata.filename` fields for a payload perhaps downloaded via HTTP.
-    if not os.path.isfile(filename):
-        return None
-
-    return get_last_modified_date(filename)
 
 
 def normalize_layout_element(
@@ -228,54 +156,6 @@ def layout_list_to_list_items(
             list_items.append(item)
 
     return list_items
-
-
-def set_element_hierarchy(
-    elements: list[Element], ruleset: dict[str, list[str]] = HIERARCHY_RULE_SET
-) -> list[Element]:
-    """Sets the parent_id for each element in the list of elements
-    based on the element's category, depth and a ruleset
-
-    """
-    stack: list[Element] = []
-    for element in elements:
-        if element.metadata.parent_id is not None:
-            continue
-        parent_id = None
-        element_category = getattr(element, "category", None)
-        element_category_depth = getattr(element.metadata, "category_depth", 0) or 0
-
-        if not element_category:
-            continue
-
-        while stack:
-            top_element: Element = stack[-1]
-            top_element_category = getattr(top_element, "category")
-            top_element_category_depth = (
-                getattr(
-                    top_element.metadata,
-                    "category_depth",
-                    0,
-                )
-                or 0
-            )
-
-            if (
-                top_element_category == element_category
-                and top_element_category_depth < element_category_depth
-            ) or (
-                top_element_category != element_category
-                and element_category in ruleset.get(top_element_category, [])
-            ):
-                parent_id = top_element.id
-                break
-
-            stack.pop()
-
-        element.metadata.parent_id = parent_id
-        stack.append(element)
-
-    return elements
 
 
 def add_element_metadata(
@@ -580,7 +460,7 @@ def _get_page_image_metadata(page: PageLayout) -> dict[str, Any]:
 # FIXME: document here can be either DocumentLayout or HTMLDocument; HTMLDocument is defined in
 # unstructured.documents.html, which imports this module so we can't import the class for type
 # hints. Moreover, those two types of documents have different lists of attributes
-# UPDATE(scanny): HTMLDocument no longer uses this function, so it can be optimized for use by
+# UPDATE(scanny): HTMLDocument no longer exists, so this function can be optimized for use by
 # DocumentLayout only.
 def document_to_element_list(
     document: DocumentLayout,
