+++ unstructured/partition/tsv.py
@@ -14,11 +14,13 @@ from unstructured.documents.elements import (
 )
 from unstructured.file_utils.filetype import add_metadata_with_filetype
 from unstructured.file_utils.model import FileType
-from unstructured.partition.common import (
+from unstructured.partition.common.common import (
     exactly_one,
+    spooled_to_bytes_io_if_needed,
+)
+from unstructured.partition.common.metadata import (
     get_last_modified_date,
     get_last_modified_date_from_file,
-    spooled_to_bytes_io_if_needed,
 )
 from unstructured.partition.lang import apply_lang_metadata
